moved total degree into own file, renamed base start system file, and other file moves /  restructures
this affects those test units which need start systems, of course.

this prepares the way for other start systems to be added at include/bertini2/system/start
