+++ core/include/bertini2/system/start_systems.hpp
@@ -1,42 +1,36 @@
 //This file is part of Bertini 2.
 //
-//system.hpp is free software: you can redistribute it and/or modify
+//start_systems.hpp is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
 //the Free Software Foundation, either version 3 of the License, or
 //(at your option) any later version.
 //
-//system.hpp is distributed in the hope that it will be useful,
+//start_systems.hpp is distributed in the hope that it will be useful,
 //but WITHOUT ANY WARRANTY; without even the implied warranty of
 //MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 //GNU General Public License for more details.
 //
 //You should have received a copy of the GNU General Public License
-//along with system.hpp.  If not, see <http://www.gnu.org/licenses/>.
+//along with start_systems.hpp.  If not, see <http://www.gnu.org/licenses/>.
 //
-// Copyright(C) 2015, 2016 by Bertini2 Development Team
+// Copyright(C) 2015 - 2017 by Bertini2 Development Team
 //
 // See <http://www.gnu.org/licenses/> for a copy of the license, 
 // as well as COPYING.  Bertini2 is provided with permitted 
 // additional terms in the b2/licenses/ directory.
 
 // individual authors of this file include:
-// Tim Hodges, Colorado State University
+// dani brake, university of notre dame
 
 /**
-\file system.hpp 
+\file b2/core/include/bertini2/system/start_systems.hpp 
 
-\brief Collects the various header files which define the Bertini2 systems.
+\brief Aggregate include for the various start systems in Bertini2.
 */
-
 
 #pragma once
 
-#include "bertini2/system/precon.hpp"
-#include "bertini2/system/start_systems.hpp"
-#include "bertini2/system/system.hpp"
-
-#endif
-
+#include "bertini2/system/start/total_degree.hpp"
