+++ core/src/system/start/total_degree.cpp
@@ -1,19 +1,19 @@
 //This file is part of Bertini 2.
 //
-//start_system.cpp is free software: you can redistribute it and/or modify
+//total_degree.cpp is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
 //the Free Software Foundation, either version 3 of the License, or
 //(at your option) any later version.
 //
-//start_system.cpp is distributed in the hope that it will be useful,
+//total_degree.cpp is distributed in the hope that it will be useful,
 //but WITHOUT ANY WARRANTY; without even the implied warranty of
 //MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 //GNU General Public License for more details.
 //
 //You should have received a copy of the GNU General Public License
-//along with start_system.cpp.  If not, see <http://www.gnu.org/licenses/>.
+//along with total_degree.cpp.  If not, see <http://www.gnu.org/licenses/>.
 //
-// Copyright(C) 2015, 2016 by Bertini2 Development Team
+// Copyright(C) 2015 - 2017 by Bertini2 Development Team
 //
 // See <http://www.gnu.org/licenses/> for a copy of the license, 
 // as well as COPYING.  Bertini2 is provided with permitted 
@@ -22,7 +22,7 @@
 // individual authors of this file include:
 // daniel brake, university of notre dame
 
-#include "bertini2/system/start_system.hpp"
+#include "bertini2/system/start/total_degree.hpp"
 
 
 BOOST_CLASS_EXPORT(bertini::start_system::TotalDegree);
