+++ core/src/system/start_base.cpp
@@ -1,41 +1,31 @@
 //This file is part of Bertini 2.
 //
-//system.hpp is free software: you can redistribute it and/or modify
+//start_base.cpp is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
 //the Free Software Foundation, either version 3 of the License, or
 //(at your option) any later version.
 //
-//system.hpp is distributed in the hope that it will be useful,
+//start_base.cpp is distributed in the hope that it will be useful,
 //but WITHOUT ANY WARRANTY; without even the implied warranty of
 //MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 //GNU General Public License for more details.
 //
 //You should have received a copy of the GNU General Public License
-//along with system.hpp.  If not, see <http://www.gnu.org/licenses/>.
+//along with start_base.cpp.  If not, see <http://www.gnu.org/licenses/>.
 //
-// Copyright(C) 2015, 2016 by Bertini2 Development Team
+// Copyright(C) 2015 - 2017 by Bertini2 Development Team
 //
 // See <http://www.gnu.org/licenses/> for a copy of the license, 
 // as well as COPYING.  Bertini2 is provided with permitted 
 // additional terms in the b2/licenses/ directory.
 
 // individual authors of this file include:
-// Tim Hodges, Colorado State University
+// daniel brake, university of notre dame
 
-/**
-\file system.hpp 
-
-\brief Collects the various header files which define the Bertini2 systems.
-*/
+#include "bertini2/system/start_base.hpp"
 
 
-#pragma once
-
-#include "bertini2/system/precon.hpp"
-#include "bertini2/system/start_systems.hpp"
-#include "bertini2/system/system.hpp"
-
-#endif
+BOOST_CLASS_EXPORT(bertini::start_system::StartSystem);
