+++ core/test/classes/start_system_test.cpp
@@ -27,7 +27,7 @@
 
 
 
-#include "bertini2/system/start_system.hpp"
+#include "bertini2/system/start_systems.hpp"
 
 using System = bertini::System;
