+++ core/test/endgames/amp_cauchy_test.cpp
@@ -34,7 +34,7 @@
 #include <iostream>
 #include <boost/test/unit_test.hpp>
 
-#include "bertini2/system/start_system.hpp"
+#include "bertini2/system/start_systems.hpp"
 #include "bertini2/num_traits.hpp"
 
 #include "bertini2/endgames/amp_cauchy.hpp"
