+++ core/test/endgames/fixed_multiple_powerseries_test.cpp
@@ -27,7 +27,7 @@
 #include <iostream>
 #include <boost/test/unit_test.hpp>
 
-#include "bertini2/system/start_system.hpp"
+#include "bertini2/system/start_systems.hpp"
 #include "bertini2/num_traits.hpp"
 
 #include "bertini2/endgames/fixed_prec_powerseries.hpp"
